% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RnaSeqTutorials-package.R
\docType{package}
\name{RnaSeqTutorial06 package}
\alias{RnaSeqTutorial06}
\alias{RnaSeqTutorial06-package}
\alias{RnaSeqTutorial06 package}
\title{RnaSeqTutorials package}
\description{
A simple description of the RnaSeqTutorial06 package
}
\section{Tutorials}{

This is the sixth in a series of tutorials
\itemize{
\item\code{06_differential_transcript_usage} a tutorial on differential transcript usage
}
}

\examples{
	\dontrun{
	learnr::run_tutorial("06_differential_transcript_usage", package = "RnaSeqTutorial06")
	}
}
\seealso{
The vignette
}
\author{
Nicolas Delhomme [aut,cre]
}
\keyword{internal}
\keyword{package}
